pyproject.toml
[project]
name = "gn-monitor"
version = "0.1.0"
description = "Get uptime and other insights about your website"
dependencies = [
    "fastapi>=0.111.1",
    "requests>=2.32.3",
    "pyopenssl>=24.2.1",
    "python-whois>=0.9.4",
    "sqlmodel>=0.0.21",
    "psycopg2>=2.9.9",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.4.0",
    "python-keycloak>=4.6.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "ruff>=0.5.5",
    "pytest>=8.3.2",
]
